enable alternate image.spec.in on image base
If you place a "_image.spec.in" file in your kiwi image sources in obs.
This alternate spec file can be found by kiwi_post_run as e.g.
/usr/src/packages/SOURCES/_image.spec.in and used for generating the
containment rpm.

For more flexibility in the images.spec.in I add the $IMAGE_NAME variable
which can be used in the images.spec.in as __IMAGE_NAME__.

An example for the obs source service containment was also added.