Include symlink to the image with a simplified name
This commit includes in the resulting docker image RPM a symlink
to the image tarball that does not include any version number, arch
and extension. The symlink is named as it is in the `name` attribute
of the `image` tag within the kiwi description file. This way the
base image version can be updated without having to also update
base image reference inside the derived image kiwi description
file.

Related to bsc#1038192