review filenames
use old.[packages|changes].$PKG_NAME.$ARCH instead, because this is all start
with old. This is just cosmetic but it is how we are doing it in the
kubik project.

Also fix the "if" clauses which were checking for the file existance
which were looking in the wrong path